Refactored out ExpandingTextFieldAndSubmitButton into it's own file
